requirements.production.txt
Django
djangorestframework
django-cors-headers
celery
redis
psycopg2-binary
python-decouple
requests
django-filter
djoser
djangorestframework-simplejwt
django-extensions
gunicorn
whitenoise
pillow
boto3
django-storages
sendgrid
icalendar
pytz

# Production WSGI server
gunicorn
gevent

# Monitoring and logging
sentry-sdk[django]
django-prometheus

# Performance
django-redis
hiredis

# Security
django-security
django-csp

# Database optimizations
psycopg2-binary

# Task scheduling
django-celery-beat